gui not displaying double as text :(

+++ GUI/GUI.h
@@ -24,7 +24,6 @@ private slots:
 	void on_btn_clear_clicked();
 	void on_btn_start_clicked();
 	void on_btn_stop_clicked();
-
 public slots:
-	void onMpcIteration(int);
+	void onMpcIteration(double);
 };
